# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Çoklu Firma Muhasebe ve Cari Takip Paneli API
servers:
  - url: /api
    description: API temel yolu
...
paths:
...
  /ekipmanlar:
    get:
      operationId: listEkipmanlar
      tags: [ekipmanlar]
      summary: Ekipmanları listele
      parameters:
        - name: sirketId
          in: query
          schema:
            type: integer
        - name: gemiId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ekipman listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ekipman"
    post:
      operationId: createEkipman
      tags: [ekipmanlar]
      summary: Yeni ekipman oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EkipmanInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ekipman"
  /ekipmanlar/{id}:
    patch:
      operationId: updateEkipman
      tags: [ekipmanlar]
      summary: Ekipmanı güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EkipmanUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ekipman"
    delete:
      operationId: deleteEkipman
      tags: [ekipmanlar]
      summary: Ekipmanı sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi

  # ── KDV Oranları ───────────────────────────────────────────────────────────
...
  /fatura-serileri:
    get:
      operationId: listFaturaSerileri
      tags: [fatura-serileri]
      summary: Fatura serilerini listele
      parameters:
        - name: sirketId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fatura serisi listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaturaSeri"
    post:
      operationId: createFaturaSeri
      tags: [fatura-serileri]
      summary: Yeni fatura serisi oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaturaSeriInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaturaSeri"
  /fatura-serileri/{id}:
    patch:
      operationId: updateFaturaSeri
      tags: [fatura-serileri]
      summary: Fatura serisini güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaturaSeriUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaturaSeri"
    delete:
      operationId: deleteFaturaSeri
      tags: [fatura-serileri]
      summary: Fatura serisini sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi

  # ── Dashboard ──────────────────────────────────────────────────────────────
  /dashboard/ozet:
    get:
      operationId: getDashboardOzet
      tags: [dashboard]
      summary: Dashboard özet verileri
      parameters:
        - name: sirketId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Özet veriler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOzet"
  /dashboard/vadesi-yaklasan:
    get:
      operationId: getVadesiYaklasanFaturalar
      tags: [dashboard]
      summary: Vadesi yaklaşan faturalar
      parameters:
        - name: sirketId
          in: query
          schema:
            type: ["integer", "null"]
        - name: gun
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Vadesi yaklaşan faturalar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fatura"
  /dashboard/son-islemler:
    get:
      operationId: getSonIslemler
      tags: [dashboard]
      summary: Son işlemler
      parameters:
        - name: sirketId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Son işlemler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SonIslemler"
  /dashboard/aylik-gelir:
    get:
      operationId: getAylikGelir
      tags: [dashboard]
      summary: Aylık gelir grafik verisi
      parameters:
        - name: sirketId
          in: query
          schema:
            type: ["integer", "null"]
        - name: yil
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Aylık gelir
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AylikGelirNoktasi"
  /dashboard/yenileme-uyarilari:
    get:
      operationId: getYenilemeyeYaklasanPlanlar
      tags: [dashboard]
      summary: Süresi dolmak üzere olan Starlink planları
      parameters:
        - name: sirketId
          in: query
          schema:
            type: ["integer", "null"]
        - name: gun
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Yenileme uyarıları
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StarlinkPlani"

  # ── Raporlar ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Fatura:
      type: object
      required: [id, sirketId, cariId, faturaNo, faturaTarihi, vadeTarihi, paraBirimi, durum, toplamTutar, kdvTutari, genelToplam, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: ["integer", "null"]
        gemiAd:
          type: ["string", "null"]
        faturaNo:
          type: string
        faturaTarihi:
          type: string
          format: date
        vadeTarihi:
          type: string
          format: date
        paraBirimi:
          type: string
        durum:
          type: string
          enum: [acik, kismi_odendi, odendi, iptal]
        toplamTutar:
          type: number
        kdvTutari:
          type: number
        genelToplam:
          type: number
        odenenTutar:
          type: number
        kalanTutar:
          type: number
        notlar:
          type: ["string", "null"]
        aciklama:
          type: ["string", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
...
    Odeme:
      type: object
      required: [id, sirketId, cariId, tip, tarih, tutar, paraBirimi, odemeYontemi, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: ["integer", "null"]
        gemiAd:
          type: ["string", "null"]
        bankaHesabiId:
          type: ["integer", "null"]
        bankaHesabiAd:
          type: ["string", "null"]
        faturaId:
          type: ["integer", "null"]
        faturaNo:
          type: ["string", "null"]
        tip:
          type: string
          enum: [tahsilat, odeme]
        tarih:
          type: string
          format: date
        tutar:
          type: number
        paraBirimi:
          type: string
        odemeYontemi:
          type: string
          enum: [banka_havalesi, eft, nakit, kredi_karti, wise, paypal, diger]
        aciklama:
          type: ["string", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
...
    StarlinkPlani:
      type: object
      required: [id, sirketId, cariId, gemiId, planAdi, baslangicTarihi, bitisTarihi, aylikUcret, paraBirimi, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: integer
        gemiAd:
          type: ["string", "null"]
        planAdi:
          type: string
        hizMbps:
          type: ["integer", "null"]
        baslangicTarihi:
          type: string
          format: date
        bitisTarihi:
          type: string
          format: date
        aylikUcret:
          type: number
        paraBirimi:
          type: string
        otomatikYenileme:
          type: boolean
        aktif:
          type: boolean
        notlar:
          type: ["string", "null"]
        kalanGun:
          type: ["integer", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
...
    Ekipman:
      type: object
      required: [id, sirketId, gemiId, tip, seriNo, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        gemiId:
          type: integer
        gemiAd:
          type: ["string", "null"]
        tip:
          type: string
        seriNo:
          type: string
        kurulumTarihi:
          type: ["string", "null"]
        garantiBitisTarihi:
          type: ["string", "null"]
        notlar:
          type: ["string", "null"]
        aktif:
          type: boolean
        olusturmaTarihi:
          type: string
          format: date-time
    EkipmanInput:
      type: object
      required: [sirketId, gemiId, tip, seriNo]
      properties:
        sirketId:
          type: integer
        gemiId:
          type: integer
        tip:
          type: string
        seriNo:
          type: string
        kurulumTarihi:
          type: string
          format: date
        garantiBitisTarihi:
          type: string
          format: date
        notlar:
          type: string
        aktif:
          type: boolean
    EkipmanUpdate:
      type: object
      properties:
        tip:
          type: string
        seriNo:
          type: string
        kurulumTarihi:
          type: string
          format: date
        garantiBitisTarihi:
          type: string
          format: date
        notlar:
          type: string
        aktif:
          type: boolean
        gemiId:
          type: integer

    # ── KDV Oranı ─────────────────────────────────────────────────────────
...
    FaturaSeri:
      type: object
      required: [id, sirketId, ad, onek, sonrakiNo, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        ad:
          type: string
        onek:
          type: string
        sonrakiNo:
          type: integer
        varsayilan:
          type: boolean
        olusturmaTarihi:
          type: string
          format: date-time
    FaturaSeriInput:
      type: object
      required: [sirketId, ad, onek]
      properties:
        sirketId:
          type: integer
        ad:
          type: string
        onek:
          type: string
        sonrakiNo:
          type: integer
          default: 1
        varsayilan:
          type: boolean
    FaturaSeriUpdate:
      type: object
      properties:
        ad:
          type: string
        onek:
          type: string
        sonrakiNo:
          type: integer
        varsayilan:
          type: boolean

    # ── Dashboard ──────────────────────────────────────────────────────────
    DashboardOzet:
      type: object
      required: [toplamAlacak, toplamTahsilat, kalanBakiye, toplamFaturaSayisi, toplamOdemeSayisi, toplamCariSayisi, vadesGecmisFaturaSayisi, vadesYaklasiyor, paraBirimiOzetleri, bankaHesapBakiyeleri]
      properties:
        toplamAlacak:
          type: number
        toplamTahsilat:
          type: number
        kalanBakiye:
          type: number
        toplamFaturaSayisi:
          type: integer
        toplamOdemeSayisi:
          type: integer
        toplamCariSayisi:
          type: integer
        vadesGecmisFaturaSayisi:
          type: integer
        vadesYaklasiyor:
          type: integer
        paraBirimiOzetleri:
          type: array
          items:
            $ref: "#/components/schemas/ParaBirimiOzet"
        bankaHesapBakiyeleri:
          type: array
          items:
            $ref: "#/components/schemas/BankaHesapBakiye"
        aktifStarlinkPlanSayisi:
          type: integer
    ParaBirimiOzet:
      type: object
      required: [paraBirimi, toplamAlacak, toplamTahsilat, kalanBakiye]
      properties:
        paraBirimi:
          type: string
        toplamAlacak:
          type: number
        toplamTahsilat:
          type: number
        kalanBakiye:
          type: number
    BankaHesapBakiye:
      type: object
      required: [hesapId, bankaAdi, hesapAdi, paraBirimi, bakiye]
      properties:
        hesapId:
          type: integer
        bankaAdi:
          type: string
        hesapAdi:
          type: string
        paraBirimi:
          type: string
        bakiye:
          type: number
    SonIslemler:
      type: object
      required: [sonFaturalar, sonOdemeler]
      properties:
        sonFaturalar:
          type: array
          items:
            $ref: "#/components/schemas/Fatura"
        sonOdemeler:
          type: array
          items:
            $ref: "#/components/schemas/Odeme"
    AylikGelirNoktasi:
      type: object
      required: [yil, ay, ayAd, toplamFatura, toplamTahsilat]
      properties:
        yil:
          type: integer
        ay:
          type: integer
        ayAd:
          type: string
        toplamFatura:
          type: number
        toplamTahsilat:
          type: number

    # ── Raporlar ──────────────────────────────────────────────────────────
...